pyproject.toml
[project]
name = "relcreator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.4"
dependencies = [
    "agno>=1.7.7",
    "fastapi[standard]>=0.116.1",
    "ollama>=0.5.1",
    "openai>=1.98.0",
    "pymupdf4llm>=0.0.27",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]

[dependency-groups]
dev = [
    "notebook>=7.4.5",
]
